+++ core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @lukekaalim/act
 
+## 5.0.0
+
+### Major Changes
+
+- 2ed02cf: Breaking! Change how useRef works, no longer compatible with React-style.
+
+  New API has `.get` and `.set` properties in order to better type ReadOnlyRef and WriteOnlyRef.
+
+  Add primitive registry to web, remove old element API and replace with "html" and "svg" element maps
+
 ## 4.3.0
 
 ### Minor Changes