+++ renderers/three/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lukekaalim/act-three
 
+## 8.1.0
+
+### Minor Changes
+
+- 2ed02cf: Breaking! Change how useRef works, no longer compatible with React-style.
+
+  New API has `.get` and `.set` properties in order to better type ReadOnlyRef and WriteOnlyRef.
+
+  Add primitive registry to web, remove old element API and replace with "html" and "svg" element maps
+
+### Patch Changes
+
+- Updated dependencies [2ed02cf]
+  - @lukekaalim/act-web@6.0.0
+  - @lukekaalim/act@5.0.0
+  - @lukekaalim/act-recon@4.0.1
+  - @lukekaalim/act-backstage@3.2.1
+
 ## 8.0.3
 
 ### Patch Changes